# perf: use numpy only in eager weight statistics