--- a/src/coffea/analysis_tools.py
+++ b/src/coffea/analysis_tools.py
@@ -212,11 +212,31 @@ class Weights:
         if self._storeIndividual:
             self._weights[name] = weight
         self.__add_variation(name, weight, weightUp, weightDown, shift)
+        if weight.size == 0:
+            dtype = weight.dtype
+            if dtype in (
+                numpy.int8,
+                numpy.int16,
+                numpy.int32,
+                numpy.int64,
+                numpy.uint8,
+                numpy.uint16,
+                numpy.uint32,
+                numpy.uint64,
+            ):
+                min = numpy.iinfo(dtype).max
+                max = numpy.iinfo(dtype).min
+            else:
+                min = numpy.inf
+                max = -numpy.inf
+        else:
+            min = weight.min()
+            max = weight.max()
         self._weightStats[name] = WeightStatistics(
             weight.sum(),
             (weight**2).sum(),
-            awkward.min(weight, mask_identity=False),
-            awkward.max(weight, mask_identity=False),
+            min,
+            max,
             weight.size,
         )
         self._names.append(name)
@@ -313,11 +333,31 @@ class Weights:
         ):
             systName = f"{name}_{modifier}"
             self.__add_variation(systName, weight, weightUp, weightDown, shift)
+        if weight.size == 0:
+            dtype = weight.dtype
+            if dtype in (
+                numpy.int8,
+                numpy.int16,
+                numpy.int32,
+                numpy.int64,
+                numpy.uint8,
+                numpy.uint16,
+                numpy.uint32,
+                numpy.uint64,
+            ):
+                min = numpy.iinfo(dtype).max
+                max = numpy.iinfo(dtype).min
+            else:
+                min = numpy.inf
+                max = -numpy.inf
+        else:
+            min = weight.min()
+            max = weight.max()
         self._weightStats[name] = WeightStatistics(
             weight.sum(),
             (weight**2).sum(),
-            awkward.min(weight, mask_identity=False),
-            awkward.max(weight, mask_identity=False),
+            min,
+            max,
             weight.size,
         )
         self._names.append(name)