Awkward is known to be much slower than plain numpy. Example
```py
In [3]: array = ak.Array(np.random.normal(size=1_000_000))

In [4]: %timeit np.min(array.layout.data)
70.6 μs ± 408 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

In [5]: %timeit ak.min(array)
3.18 ms ± 20.3 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```
So I'd rather follow the pattern and use plain numpy in eager weights/packed selection.